python-service/requirements.txt
roma-dspy
fastapi
uvicorn[standard]
python-dotenv
pydantic
